bundle: Rename JSON protocol related symbols
This is a preparation for moving bundle.Args and its friends to
jsonprotocol package. Several symbols in args.go are renamed to
add "Bundle" prefix.

This patch fails on pre-upload hooks due to type names having their
bundle name as a prefix, but I ignored them since they are soon
resolved by a change to move those types to jsonprotocol package.

This is a pure rename change.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I9d2f69810629ffad0bea270eb4fa606f75a7ccbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2804860
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>