bundle: Pass WaitUntilReady value via Context
bundle.Delegate provides two hooks to be called on the beginning of
a run: RunHook and Ready. But in fact both hooks are internally merged
to a single run hook when Delegate is converted to staticConfig.

An odd thing is that user-written run hooks can't see values of
-waituntilready, while ready hooks can. Actually ready hooks cheats by
capturing jsonprotocol.BundleArgs, which is available as a local
variable in newArgsAndStaticConfig.

In order to capture the variable, we have to create BundleArgs at the
same time as staticConfig in newArgsAndStaticConfig. This makes it
difficult to get rid of jsonprotocol dependencies.

This patch solves the problem by making -waituntilready value available
to ready hooks via formal private API using Context. New context
functions are private to the framework and should never be exposed to
user code.

In the longer term, we should drop or replace -waituntilready. Meanwhile
This change eases isolating jsonprotocol dependencies.

This is a refactoring change. No functional change is expected.

BUG=chromium:1128259
BUG=chromium:1195719
TEST=fast_build.sh -T

Change-Id: Ied090f61b807e7330367f29d72d07dbb637399f7