testing: Record registration errors in Registry object
As of today, while we can instantiate multiple testing.Registry objects,
registration errors are recorded in a global array. This makes it hard
to write unit tests that work with multiple fake bundles (e.g. runners).

This patch enhances Registry object to record errors as well.

This is a refactoring change. No functional change is expected.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Ic921b9957973b02d78a43216fc5fbed16d48867f