bundle: Always call newStaticConfig to construct staticConfig
In an upcoming change we want to add testing.Registry to staticConfig.
To ensure to always set the field in unit tests, rewrite all wild struct
initialization of staticConfig with a call to newStaticConfig, with
which we can catch missing arguments.

This is a pure refactoring change.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I2c8097393aae50532e536b27c856842f37758b96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2807540
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>