tast: Remove dependency to infra/proto message type.
The device.Config message type in infra/proto protobuf is used by the
infra code internally, but it has also been used by Tast for
communication between processes.
It was once needed for infra-side test scheduling plan, but it is
changed to a different format. Therefore Tast no longer need to use it.
In order to remove that dependency, protocol.DeviceInfo struct is
introduced to store all device information used by Tast but not exist
in config.api.HardwareFeatures.

BUG=b:187792694
TEST=fast_build.sh -t chromiumos/tast/internal/runner chromiumos/tast/testing/hwdep chromiumos/tast/internal/bundle chromiumos/tast/cmd/tast/internal/run/runnerclient

Change-Id: I3342c7c9d45b7a40d3a40524ad2d60486f9072f0