bundle: Remove JSON protocol deps from runTests
runTests is one of the most important functions in test bundles that
drives tests. In later changes we want to call the function from gRPC
servers as well.

This patch removes dependencies to JSON protocol from runTests by
replacing jsonprotocol.BundleArgs with protocol.RunConfig.

This is a refactoring change. No functional change is expected.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Iddd072e1bf75c456f4d66d44f3886f519d550156