bundle: Remove bundleType
Instead of statically specifying whether a bundle is local or remote,
determine it by whether protocol.RunConfig.RemoteTestConfig is set or
not. This allows us to step towards general bundle implementations.

Note that -help output is slightly changed to drop references to
test_runner. But it is rarely seen by users.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I37123e088b3d060dafad4d5bcc0873ec064e124a