bundle: Stop passing TestInstance to runTests
runTests accepts []*TestInstance, but it can be computed by the pair of
protocol.RunConfig and bundle.StaticConfig. To avoid ambiguity, stop
passing []*TestInstance to runTests and instead compute a list of test
instances within the function.

This is a refactoring change. No functional change is expected.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Ie3f7ca0a267e4b83569047a2ba16ab1343c3b811
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2831692
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>