bundle: Do not fail when no tests are to be executed
Currently a test bundle exits abnormally when it is requested to run
no test. However running an empty set of tests is technically a valid
request, so specially handling the case harms generality.

This patch removes the logic to handle this case specially.

Fortunately the code path is not triggered today because test runners
do not run test bundles with no hit tests, so this patch should not
change observable behavior.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Id0c04c75eb8311f81003a183d3b98c71f7803e64