rpc: Ensure cleanup functions are called
When a gRPC server is interrupted for some reasons, it can exit
before running deferred cleanup functions. This is not desirable
as we move the framework to fully rely on gRPC.

This patch ensures cleanup on the following cases:

1. When stdout is closed. This is a very common case for binaries
   executed on an SSH session without tty, e.g. local_test_runner.
   By default, writing data to closed stdout causes SIGPIPE to be
   delivered to terminate the process. This patch updates the server
   to just ignore SIGPIPE. gRPC framework stops serving once it
   detects a connection error and closes the connection.

2. When SIGINT/SIGTERM is sent to the process. This is a case for
   binaries executed on a session with tty, e.g. remote_test_runner.
   By default, processes are terminated immediately. This patch
   installs a handler for those signals and stops the server.

In both cases, it is important to explicitly wait for all active gRPC
method calls to finish. gRPC does not guarantee it upon completion of
grpc.Server.Serve.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I0563a4a657b90745dd917732ee3444d29cd5fe51