rpc: Enhance DialExec to support starting a new session
If newSession argument is true, rpc.DialExec creates a new session
to start a subprocess and kill all processes in the session on closing
the connection.

Test runners use this flag to ensure no subprocesses remain in RunTests
gRPC call after it is interrupted and test bundles are killed. For
consistency the same behavior is applied to ListEntities gRPC calls, but
it is mostly no-op since test bundles do not spawn subprocesses in the
case.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: I67fcc1476d67cd844b068b819e6c935af693feaf