rpc: Split Client types for different connection types
In the following changes we need to obtain PID of a gRPC server process,
but the current API does not allow it because the same type (Client) is
returned from both DialSSH and DialExec.

This patch updates dial functions to return distinct types so that they
can provide different sets of methods.

  DialExec  -> ExecClient
  DialSSH   -> SSHClient
  NewClient -> GenericClient

Public API for user code is preserved.

BUG=chromium:1128259
TEST=fast_build.sh -T

Change-Id: Ifefb2ae81d075e5f9fbf6b5a4bc5f3a4447bfae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2856396
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>