run: Introduce genericexec package
Today we use the runnerCmd interface in runner_cmd.go as an abstraction
of external command execution on the local machine and over the SSH.
However its interface is not flexible enough to run gRPC connections
over its stdin/stdout.

This patch introduces a new package "genericexec" to provide a more
well-designed API with unit test coverage.

The new package is used nowhere now. The next patch replaces runnerCmd
with it.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: I9469527732b7f6b616537864149eae82234489e9