run: Replace State fields with DUTInfo
This patch replaces several fields in the State struct that are filled
in GetDUTInfo with a single protobuf message DUTInfo.

We still use the same JSON-based protocol to send DUTInfo from
local_test_runner to Tast CLI.

This is a refactoring change. No functional change is expected.

BUG=b:187793617
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run $DUT meta.LocalPass meta.RemotePass

Change-Id: I007cd54348fc3913d93b59ede15ca64811ca0179
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2905153
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>