Add a "chromeless" softwaredep.
In https://crrev.com/c/2889934, we wish to distinguish between systems
that do not have chrome installed and those that do (but run the test on
both). To do so, we add a softwaredep for "chromeless", and will (in a
future CL) create one subtest each for systems with / without chrome.

BUG=b:186688190
TEST=CQ

Change-Id: I61722e2a964ea83dac76ef7429c0235b69ac34bf