tast: Add CPU, SoC and battery fields to DeviceInfo.
These data has been stored to device.Config struct, but we will migrate
it to the tast-internal struct for removing the dependency to
infra/proto/go/device.

HardwareFeatures is only used as an internal variable in Tast framework,
but not used for communication between processes.
Therefore the schema can be modified breaking the protobuf wire format
compatibility.

There may be version discrepancy between DUT and host. This change keeps
compatibility for the both cases:
1. When the host-side program is old:
  - DUT populates both the old and new message. The server inteprets
  the old one.
2. When the DUT-side program is old:
  - Server-side code accept both new and old messages and translates it.

BUG=b:187792694
TEST=TBD

Change-Id: Ifcf07ca48496d9e6796a977f0a857f2745fce52e