fakebundle: Add InstallAt
Install function creates a temporary directory automatically. This is
convenient in some cases but not in other cases where we want to
install fake bundles at a specific directory.

This patch adds InstallAt function that allows callers to specify a
directory to which fake bundles are installed.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: I7641dbe0f1c1013d04f13fcf9f36ac307c1e1497
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2919428
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>