devserver: Do not wait for active connections on close
Currently the ephemeral devserver waits for all active connections
before shutting down, but this is not really needed as Tast knows that
all connections can be aborted when it is shutting down the ephemeral
devserver after running test bundles.

This is essentially a one-line change, but rewrite several other places
as context.Context is no longer necessary.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: I8106a0d2098fb65f3216537ecdad74e40688aaee