runtest: Support gRPC mode in fake test runners
This is a part of patch series to introduce the new "runtest" package
that provides common testing environment for Tast CLI unit tests.

This patch adds gRPC mode support in the fakerunner package. The
implementation strategy is the same as JSON mode, that is, to call
into the real implementation of test runners.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: I4a6045739ee21b2310364b3660e5ecb90df43df9