runtest: Introduce fakesshserver package
This is a part of patch series to introduce the new "runtest" package
that provides common testing environment for Tast CLI unit tests.

Currently Tast CLI unit tests use c/t/cmd/tast/internal/run/fakerunner
package to create fake test runners. But the package has several
limitations:

1. The package is strongly tied to the JSON protocol. This prevent us
   from running unit tests against gRPC protocol mode.

2. The package has limited support of remote tests. Due to this, Tast
   CLI ended up having sparse test coverage for remote testing in
   general. This becomes more problematic when we change Tast CLI to
   drive local test bundles via remote test bundles, which is planned
   for the general remote fixture support.

Thus this patch series introduce a new package runtest to solve these
limitations. The existing fakerunner package is deprecated, and will
be deleted once all users are migrated to the new package.

This patch is a first step that introduces a new internal package
that implements a fake SSH server. It is designed to be flexible
enough to simulate test runners and other arbitrary commands.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: I06f8385ffe6fc50ad4bfdcaaa342cc31f374b973