tast-lint: Allow linting from git subdirectories
At present execution of tast-lint only works on Git root directory.
So if it is executed from any subdirectories where ".git" is not
available, the main goroutine panics.

In general, tast-lint checks linting error in the specified commit hash
and tries to fix it. So it should have the flexibility to check linting
from any subdirectory, not only from the git root.

BUG=None
TEST=./fast_build.sh -T

Change-Id: I5d57f7b14579154736a8322376863c31c1160ae2