tast: Allow running tast-lint from arbitrary directories
At present tast-lint only works on Git root directory. If it is executed from any subdirectories other than the Git root where .git is not available, the main goroutine panics.

In order to allow it running on any subdirectory, the patch manually navigates to the root directory of its immediate parent repo.

BUG=None
TEST=./fast_build.sh -T

Change-Id: I5d57f7b14579154736a8322376863c31c1160ae2