tast-lint: Allow linting from git subdirectories
At present execution of tast-lint only works on Git root directory.
So if it is executed from any subdirectories where ".git" is not
available, the main goroutine panics.

Also, there was an existing BUG in tast-lint. Currently, passing go
file with an absolute path as an argument to tast-lint panics the
main goroutine. For eg.

tast-lint /chromeos/src/platform/tast/src/chromiumos/tast/cmd/tast/main.go

doesn't works.

In general, tast-lint checks linting error in the specified commit hash
and tries to fix it. So it should have the flexibility to check linting
from any subdirectory, not only from the git root.

Also, if directory has been changed internally, this should not affect
the usage of tast-lint in any way. On the contrary, it should support
sophisticate `tast-lint ./main.go` kind of statement from
subdirectories.

BUG=None
TEST=./fast_build.sh -T

Change-Id: I5d57f7b14579154736a8322376863c31c1160ae2