tast-lint: Efficiently handles fmt.Errorf in tast-lint if errors package is already imported
At present, using go ast, tast-lint checks for "errors" identifier node to determine if "chromiumos/tast/errors" package has been used or not. If it is used, during "tast-lint -fix" for fmt.Errorf it simply ignores that.

This CL, instead of checking identifier node, scans through all the imports and analyze if "errors" package is imported with also suitable import alias. If impoted, it update the fmt.Errorf with <scan_alias>.Errorf. It also takes care of avoiding namespace collision for dot import.

BUG=chromium:1142132, b:187794245
TEST=./fast_build.sh -T

Change-Id: I547b810734b5b477d758063a234e0924ba543845