tast: Dump goroutines to stderr when the process gets SIGTERM
This allows us to see what Tast was doing when it hits timeouts.

Also the messages are now written to stderr so that they are not caught
by the logic in tast.py to read the last line of error messages.

Note that this change drops SIGKILL from the list of registered signals
since Linux processes cannot handle SIGKILL anyway.

BUG=b:189332919
TEST=fast_build.sh
TEST=timeout 10 ~/go/bin/tast run $DUT meta.LocalFreeze

Change-Id: I4622cb477089d2c2c8101d3670ff37fca950ac23