testing: Assign name to registry
{Test,Fixture}Instance.EntityProto returns a protobuf representation
of them, and it contains "Bundle" field indicating which test bundle
it belongs to.

In production, there is only one meaningful registry (aka "the global
registry"), thus it is sufficient to return filepath.Base(os.Args[0])
as a bundle name. However, in unit tests, there can be several
registries at a time, so the same method cannot be used to obtain a
test bundle name.

This patch solves the problem by assigning a name to a registry and
associating TestInstance/FixtureInstance to the registry on adding
them to it (AddTestInstance/AddFixtureInstance). The global registry
will get the name as filepath.Base(os.Args[0]), but unit tests can
create new registries with arbitrary names.

fakebundle/runtest packages are also updated to make use of registry
names.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: I7d3e3d49121d93e1ecb2b162d3923c5842cdc828