tast: Expose DUTFeatures in testing.State
Allow tests and preconditions to see the DUTFeatures proto, by adding a
DUTFeatures() method to entityMixin.

It is sometimes useful, particularly in Preconditions, to be able to
inspect the DUTFeatures and behave differently. The case I am thinking
of at the moment is the difference between a DUT with a ChromeEC vs a
WilcoEC when cleaning up in the firmware boot mode precondition.

BUG=None
TEST=None

Change-Id: I9293ae0abcccff40b78d801feac78fd218659067