tast: Allow dereference symlinks in GetFile()
Getting a symlink was only downloading the link itself, not the file it
pointed to. Adding the --dereference flag to the tar command makes
GetFile work as expected for a file (you get the file the symlink points
to), and when calling GetFile on a directory containing a symlink you
continue to get the symlink.

BUG=b:185998785
TEST=Called GetFile on /var/log/servod_9999/latest.DEBUG (got file) and
/var/log/servod_9999 (latest.* remained as symlinks)

Change-Id: I903acc68ed977991dc2bb86f7b5318d1506ad1e3