tast: Support for Closers
A utility has been introduced to help in closing the acquired resources
through a single interface instead of calling defer multiple times.

It is useful in following scenarios

1. Shortening the context for post test cleanup is a norm in tast, but
it's highly unlikely that the all cleanup methods will be available
during the start of the test.

2. Calling the required defers with a timeout unnecessarily  increases
number of lines. An adhoc timeout is generally used by the programmer,
not the best approach.

Instead, collect all the defer closers, store it in a single slice,
shorten the context for cleanup, perform the cleanup through a defer
block.

BUG=None
TEST=None

Change-Id: I68f4ad9d02ba87928a1afb0dec98110a5b9ade71