tast: add "no_ath10k_4_4" dependency
There are some driver features (esp. those covered in b/138406224) that
weren't backported to the ar10k copy of ath10k found in kernel 4.4. Some WiFi
tests may need to be skipped for such systems.

Technically, this would be a cross section of a Harware (the Qualcomm
QCA6174A WiFi chip) and Software (kernel 4.4) dependency, but because
board:scarlet only used 1 WiFi chip (and because no other platform used
this chip on kernel 4.4), we can get away with exclusively a
SoftwareDep.

BUG=b:160281713
TEST=run new tests using this dep, on board:scarlet and on board:grunt

Change-Id: Ib97a236b8f0f8850fc53a377d7da1c1cf73f1813