tast: Add support for hwdep.Microphone() and hwdep.Speaker().
Report Microphone() as satisfied if there is at least one lid microphone
or base microphone, using the presence of an INTERNAL_MIC or FRONT_MIC
for lid microphone, and REAR_MIC for base microphone, in
cras_test_client output. This does not distinguish between one and more
than one since that's not easily discoverable and isn't currently
required.

Report Speaker() as satisfied if a speaker_amplifier field is present.
The name string within it is not used as that information is not readily
available and that detail is not needed for this particular hwdep.
Populate the speaker_amplifier field when a INTERNAL_SPEAKER is present
in cras_test_client output.

BUG=b:189921066
TEST=fast_build.sh -T

Cq-Depend: chromium:3015906
Change-Id: Ibeb617e4efac649bf69076180d0e8ea3961ed447