tast: Add support for hwdep.Microphone() and hwdep.Speaker().
The topology proto doesn't quite line up with the requirements for these
hwdeps (presence, rather than details; microphones aren't directly
covered at all), so use the device form factor to determine speaker and
microphone presence (present on all but Chromebox, Chromebit and unknown
form factor). This does have a limitation that speaker and microphone
may be included in a Chromebox, but that hasn't happened thus far, so we
can defer on this.

One additional complication is cros_config reports the form factor as
CHROMEBOOK (and occasionally REFERENCE) for clamshell, convertible,
detachable and chromeslate. Map these to convertible until cros_config
is populated with the precise Boxster form factor. For microphone and
speaker presence, this granularity is sufficient.

BUG=b:189921066
TEST=fast_build.sh -T

Change-Id: Ibeb617e4efac649bf69076180d0e8ea3961ed447