tast: DUT.WaitUnreachable connects if needed
Change WaitUnreachable() to attempt to connect instead of failing when
the ssh connection is not connected. If it fails to connect, fantastic,
it's already unreachable and we don't have to wait.

Add logging when waiting, but remove the logging on every connect.

BUG=None
TEST=ran all firmware_smoke tests.

Change-Id: Iafa4afe3325c6851ff60aa63de28ec2b27d426e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3017074
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>