Ignore -maybemissingvars
This a one-off change to fix CI for R90 branch.
The CI provides -maybemissingvars but Tast on R90 branch
doesn't know about it and it caused an error.
Unitl the long term solution on infra is implemented, this change
makes Tast on R90 branch accept the flag, just ignoring it.

BUG=b:191779650
TEST=manually gave the flag, and it worked

Change-Id: If57a48bed99710f0079f73a37a9789efc0607454
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3062142
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>