run: Introduce BundleEntity
In later changes, we have to tell a bundle that an entity belongs to
on listing and running tests in Tast CLI.

Though protocol.Entity (and protocol.ResolvedEntity) contains a bundle
name in its LegacyData field, it exists just for compatibility reasons
and will be removed in the future.

Thus this change introduces BundleEntity struct which is essentially
a pair of protocol.ResolvedEntity and a bundle name, and rewrites most
references of protocol.ResolvedEntity with BundleEntity.

For now, we define BundleEntity at a new package run/driverdata.
Ideally we want to define it within the driver package, but it is not
possible today due to circular imports caused by its usage in
config.State. We got one more reason to kill State struct.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: I3f4b853e5edc550c99564417a6ee35ad8a3f3a4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3215942
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>