bundle: extract a function for recursive listing.
This is part 12 of the patches to support full remote fixtures.

The function for recursively listing entities is extracted.
The function will be used for recursively running tests; we will retrieve
entities in the target machine and use them for planning tests running
order.

BUG=b:187957164
TEST=fast_build.sh -T

Change-Id: I4f6b3969c2de1c5340b54f14635cf7ea1a815c6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3220273
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>