planner: split output package
This is part of the patches to support full remote fixtures.

In this change, we extract out functionalities for outputting
entity results from planner package into planner/internal/output
package.

In a later change we will extract fixtures package from planner too
and this change is a necessary preparation for that, because
the fixtures package will use the output package.

No functional changes expected.

BUG=b:187957164
TEST=fast_build.sh -T

Change-Id: I94ed33f51bcb5f87a5302d12dfb294826f398927
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3241584
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>