planner: Split fixture package
This is part of the patches to support full remote fixtures.

In this patch and following several changes, we are going to implement
a fixture stack across machines.
As a first step, this patch moves existing fixtures related code to
newly created fixtures package.
Along the change, we rename FixtureStack to BundleStack, and also
put a dedicated config for BundleStack as fixtures.Config separately
from planner.Config. fixtures.Config is a subset of planner.Config,
which is just enough for BundleStack to work.
Common utilities between fixtures and planner are now put
in another new package named entity.

BUG=b:187957164
TEST=fast_build.sh -T

Change-Id: I7a898f2482dfcfd2ab6e8615b01217b5a7bc84e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3241586
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>