fixtures: Implement combined stack and external stack
This is part of the patches to support full external fixtures.

This patch adds a combined stack struct which combines two fixtures
and shows it as if it's a single complete fixture stack.

This patch also adds a external stack struct which allows local bundles
to operate external fixture stack without knowing the fixture stack
protocol between bundles.

Planner in the local bundle will use a combined stack, setting a
external stack as a parent and a bundle stack as a child.

Currently we don't have unit tests, but succeeding planner update will
cover them.

BUG=b:187957164
TEST=fast_build.sh -T

Change-Id: If50d60f2fef1b315426c29f041ac64e840639f42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3244704
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>