planner: Reform config using protocol messages
This is part of the patches to support full remote fixtures.

Some of the fields in Config can be refactored using
messages in protocol package. It makes Config have less number of fields
and be more structured.
It's a clean up before adding a few another fields to Config in a later
patch.

BUG=b:187957164
TEST=fast_build.sh -T

Change-Id: Ibdf1892cba0566556e889080ef7062040ba951c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3246776
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>