linuxssh: Add GetAndDeleteFilesInDir
This is part of the patches to support full remote fixtures.

We will use this function to pull output files in a local bundle to
a remote bundle. We have been using GetAndDeleteFile for this purpose,
but we cannot use it after remote fixture support is added, which is
because GetAndDeleteFile removes the destination directory no matter what.
Remote fixture's PreTest and PostTest run before test finishes,
meaning the output directory might be already populated by PreTest
and PostTest when a test has finished, so we must keep existing files.
GetAndDeleteFilesInDir does exatly that; pulls the files on host without
deleting existing files in the destination directory.

BUG=b:187957164
TEST=fast_build.sh -T

Change-Id: Id4a8683b9e7b7844381d9a86df1af39ec933e561
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3246777
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>