runner: Redirect JSON-based protocol methods to gRPC service
This patch moves the implementation of all test runner RPC methods to
the gRPC server, and modifies JSON-based protocol handlers to delegate
to it with a local compatibility server.

You'll notice that the body of GetDUTInfo/GetSysInfoState/CollectSysInfo
were moved to the crosbundle and their implementations in the gRPC
service are very thin, while DownloadPrivateBundles is not. This is
intentional difference because DownloadPrivateBundles is a RPC method
to be called before test bundles are ready and thus it cannot be moved
to test bundles.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: Ib8aa8e62858388c81b359c4a721ce78af1bdc071
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3263614
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>