Move software feature definitions to the crosbundle package
Eventually all Chrome OS specific logic should be removed from the
framework. In fact, hard-coding these logic in the framework makes
it difficult to write certain kind of unit tests already.

This patch is part of the series of patches to move Chrome OS
specific logic to the crosbundle package, which paves the way to
remove them from the framework.

This is a pure code move. No functional change is expected.

BUG=b:187793617
TEST=fast_build.sh -T

Change-Id: I2f7cccd89c5ed740b949c23769c594bcef05153e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3256982
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>