internal/testing: Introduce FixtTestEntityRoot
This is part of the patches to support full remote fixtures.

This patch adds testing.FixtTestEntityRoot, which is designed to produce
a FixtTestState passed to PreTest and PostTest.

Because PreTest and PostTest are run as a part of the test, the root
objects producing states for them must share the error status, so that
state's HasError method works as intended. For this purpose,
testing.EntityCondition is introduced.

BUG=b:187957164
TEST=fast_build.sh -T

Change-Id: Ib81e830df1d5857f01471510ca98f79c942d69fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3263616
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>