tast: Update Sharding to Have Fixed Intervals
Sharding of tests has been flaky on real boards due to likelihood
that "Skippable" tests is not consistent between runs.
This change ensures that all shards have the same fixed distro of
tests regardless of whether the tests are skippable or not.
The downside is that runnable tests may not be evenly distributed
between shards.

BUG=b:202689823
TEST=~/trunk/src/platform/tast/fast_build.sh -T

Change-Id: Iab6a0c6dae64bd193d4e93ecc16eabd7edd1a6bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3276619
Tested-by: Eric Lok <lokeric@google.com>
Commit-Queue: Eric Lok <lokeric@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>