runnerclient: Fix message timeout configuration
MsgTimeout specifies the maximum duration for which DUTs can be
unresponsive. Its default value is 1 minute, which allows some local
tests to briefly make the DUT go offline.

In JSON-based protocol, test runners and test bundles send heartbeat
messages every second. Thus we can consider that the DUT is down if
we don't receive any message for MsgTimeout.

In gRPC-based protocol, heartbeat messages are not defined. Thus it is
not enough to consider that the DUT is down even if no messages are
received for MsgTimeout.

For example, if a test doesn't emit any log for 55 seconds and then the
DUT goes offline, Tast CLI aborts only 5 seconds after it goes offline,
which is too early.

This patch fixes the issue by reconfiguring the keepalive settings to
allow longer timeouts for ping responses. Ideally we want to send pings
more frequently, but due to gRPC's keepalive implementation details,
we cannot send ping packets frequently if we want longer ping timeouts,
so we end up waiting up to 2 minutes in the worst case.

BUG=b:187793617
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=TAST_EXP_USE_GRPC=1 ~/go/bin/tast run $DUT network.DiagPassing.video_conferencing

Change-Id: I66bd05602032bf8c7903631fd64a3b8e7c4931a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3288851
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>