target: Remove dependency to config package
This is part of the changes to extract minidriver for bundles to use.

This patch updates constructors in the target package to take only
necessary information through its own config structs. This makes
target a standalone package which doesn't depend on outer config
package, making it possible to move it outside run/ package.

BUG=b:187957164
TEST=./fast_build.sh -T

Change-Id: I5741202d2bf75934a61055f3ace8bb3984ebf5e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3312993
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>