Enable attaching the debugger to the new tast driver.
The debugger is currently not working because the files the code was stored within have been deleted and replaced with the new driver.

This will allow attaching the debugger to local and remote bundles with the new driver, and disable attaching it to the test runners, as that feature doesn't work with the new driver.

BUG=b:199006572

TEST=Confirmed it was able to connect locally.

Change-Id: Ia1dee45b844069e803b30fdde537b4a3c3fbc8f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3312986
Auto-Submit: Matt Stark <msta@google.com>
Tested-by: Matt Stark <msta@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Matt Stark <msta@google.com>