minidriver: Extract functionalities to run external tests
This is part of the changes to extract minidriver for bundles to use.

This patch extracts the function to run tests on DUT into minidriver
package.
Currently driver is the only user of this function, but when we make
remote tests invoke local tests, to support remote fixture, bundle code
would make sure of this function.

No functional changes are expected.

BUG=b:187957164
TEST=./fast_build.sh -T

Change-Id: I706211d2ec5de582d16c8fd015b4f988c48430d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3322076
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>