run: Move packages minidriver will use (part 3)
This is part of the changes to extract minidriver for bundles to use

This package moves reporting and resultsjson package.
This patch cannot be merged with the part 1 patch, because the moved
packages depend on the packages moved in the part 2.

This is a pure code move, and no functional changes are expected.

BUG=b:187957164
TEST=./fast_build.sh -T

Change-Id: I25148fbdba268544deb69175ee7a6a2b5c3b94e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3373008
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Keigo Oka <oka@chromium.org>