driver: Move packages minidriver will use (part 4/4)
This is part of the changes to extract minidriver for bundles to use

This patch moves processor package.
This patch cannot be merged with the patch part 2, because the moved
packages depend on the packages moved in the part 3.

This is a pure code move, and no functional changes are expected.

BUG=b:187957164
TEST=./fast_build.sh -T

Change-Id: I00e51c601cc07214b9c60fa8b945edf809fddf3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3373009
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Keigo Oka <oka@chromium.org>